# api.yaml
swagger: '2.0'
info:
  title: API микросервиса авторизации
  version: 1.0.0
paths:
...
  /auth/v1/settings/motivation-config:
    get:
      tags:
        - MotivationCheckSettings
      description: Запрос получения настроек программы мотивации
      operationId: getMotivationCheckSettings
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/motivationSettings'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
    post:
      tags:
        - MotivationCheckSettings
      description: Запрос обновления настроек программы мотивации
      operationId: updateMotivationCheckSettings
      produces:
        - application/json
      parameters:
        - in: body
          name: updateMotivationSettingsRq
          description: Тело запроса обновления настроек программы мотивации
          required: true
          schema:
            $ref: '#/definitions/motivationSettings'
      responses:
        200:
          description: ОК
        400:
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/errorData'
        403:
          description: Доступ запрещен
        500:
          description: Internal server error
  /auth/v1/motivation/hide-window:
    get:
      tags:
        - Motivation
      description: Запрос признака необходимости отображения окна мотивации
      operationId: getMotivationWindowHide
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            type: boolean
            description: >
              признак необходимости отображения окна мотивации:
                true: скрывать окно
                false: отобразить окно
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation:
    get:
      tags:
        - Motivation
      description: Запрос программы мотивации для текущего пользователя
      operationId: getMotivation
      produces:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/motivationProgramUser'
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
    post:
      tags:
        - Motivation
      description: Запрос на создание или редактирование программы мотивации для текущего пользователя
      operationId: postMotivation
      parameters:
        - in: body
          name: motivationProgramUserModify
          description: Запрос на создания или редактирование программы мотивации для текущего пользователя
          required: true
          schema:
            $ref: '#/definitions/motivationProgramUserModify'
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:
          description: ОК
          schema:
            $ref: '#/definitions/motivationProgramUser'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        404:
          description: Not Found
        406:
          description: Невозможно внести изменения в мотивацию в проверенном статусе
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/print:
    get:
      tags:
        - Motivation
      description: Получить данные печатной формы для мотивации.
      operationId: printMotivation
      produces:
        - application/vnd.openxmlformats-officedocument.wordprocessingml.document
      responses:
        200:
          description: OK
          schema:
            type: string
            format: byte
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/{motivationDocumentTypes}/attach-document:
    post:
      tags:
        - Motivation
      description: Прикрепить файл к мотивации.
      operationId: uploadMotivationAttachment
      parameters:
        - in: formData
          name: content
          type: file
          description: Прикрепляемый файл
        - in: path
          name: motivationDocumentTypes
          type: string
          description: >
            Типы документов мотивации:
              PASSPORT - Паспорт
              CHECK_FORM - Форма, подтвержденная подписью
          required: true
      responses:
        200:
          description: ОК
          schema:
            type: string
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        403:
          description: Невозможно прикрепить документ
        423:
          description: Функциональность отключена
        404:
          description: Not Found
        500:
          description: Internal server error
    get:
      tags:
        - Motivation
      description: Получить файл, прикрепленный к мотивации.
      operationId: getMotivationAttachment
      parameters:
        - in: path
          name: motivationDocumentTypes
          type: string
          description: >
            Типы документов мотивации:
              PASSPORT - Паспорт
              CHECK_FORM - Форма, подтвержденная подписью
          required: true
      produces:
        - application/octet-stream
      responses:
        200:
          description: OK
          schema:
            description: Контент файла
            type: string
            format: byte
        400:
          description: Bad request
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/{motivationId}/edit:
    post:
      tags:
        - Motivation
      description: Редактирование настроек программы мотивации пользователя
      operationId: motivationByMotivationIdEdit
      parameters:
        - in: path
          name: motivationId
          type: integer
          format: int64
          required: true
          description: Идентификатор программы мотивации
        - in: body
          name: updateMotivationSettingsRq
          description: Тело запроса обновления настроек программы мотивации
          required: true
          schema:
            $ref: '#/definitions/motivationSettingsEditAdmin'
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: ОК
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/errorModel'
        403:
          description: Forbidden
        404:
          description: Not Found
        406:
          description: Не все необходимые данные присутствуют для перевода в статус корректного
          schema:
            $ref: '#/definitions/errorModel'
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/{motivationId}/get-document/{motivationDocumentTypes}:
    get:
      tags:
        - Motivation
      description: Получить файл, прикрепленный к мотивации.
      operationId: getMotivationAttachmentById
      parameters:
        - in: path
          type: integer
          format: int64
          required: true
          name: motivationId
          description: Идентификатор обращения клиента
        - in: path
          name: motivationDocumentTypes
          type: string
          description: >
            Типы документов мотивации:
             PASSPORT - Паспорт
             CHECK_FORM - Форма, подтвержденная подписью
          required: true
      produces:
        - application/octet-stream
      responses:
        200:
          description: OK
          schema:
            description: Контент файла
            type: string
            format: byte
        400:
          description: Bad request
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/{userId}/get-latest:
    get:
      tags:
        - Motivation
      description: Получить активную мотивацию для пользователя
      operationId: getUserMotivationLatest
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          name: userId
          description: Идентификатор пользователя
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/motivationProgramAdmin'
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/{userId}/get-hystory-list:
    get:
      tags:
        - Motivation
      description: Получить список мотиваций для пользователя
      operationId: getUserMotivationsList
      produces:
        - application/json
      parameters:
        - in: path
          type: integer
          format: int64
          name: userId
          description: Идентификатор пользователя
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/motivationProgramAdminList'
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
  /auth/v1/motivation/report:
    get:
      tags:
        - Motivation
      description: Вывод отчета программ мотивации.
      operationId: getXlsxReport
      produces:
        - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      parameters:
        - in: query
          name: startDate
          type: string
          format: date
          description: Дата начала формирования отчета
          required: false
        - in: query
          name: endDate
          type: string
          format: date
          description: Дата окончания формирования отчета
          required: false
      responses:
        200:
          description: OK
          schema:
            type: string
            format: byte
        403:
          description: Forbidden
        404:
          description: Not Found
        423:
          description: Функциональность отключена
        500:
          description: Internal server error
definitions:
...
  errorData:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            fieldName:
              type: string
              description: Поле, к которому относится ошибка
            errorMessage:
              type: string
              description: Сообщение об ошибке
...
  motivationSettings:
    type: object
    required:
      - isEnabled
    properties:
      isEnabled:
        type: boolean
        description: Признак, показывающий активен ли модуль или нет
      expireTime:
        type: integer
        description: Срок действия мотиваций в месяцах
  motivationProgramUser:
    type: object
    description: Программа мотивации
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор мотивации
      accountNumber:
        type: string
        description: Номер счета банка РСХБ
      bikCode:
        type: string
        description: БИК банка
      bankName:  
        type: string
        description: Наименование филиала банка
      inn:  
        type: string
        description: ИНН
      registrationAddress:  
        type: string
        description: Адрес регистрации, в случае отсутствия ИНН
      index:  
        type: string
        description: Индекс, в случае отсутствия ИНН
      comment:
        type: string
        description: комментарий
      motivationCorrectStatus:
        $ref: '#/definitions/motivationCorrectStatus'
        description: Статус корректности мотивации
      documentName:
        type: string
        description: Имя прикрепленного документа с подтверждением участия
      passportName:
        type: string
        description: Паспорт участника программы мотивации
      startDate:
        type: string
        description: Дата начала действия програмы мотивации в формате ISO-8601 (ГГГГ-ММ-ДД)
  motivationProgramUserModify:
    type: object
    description: Модель создания или модификации программы мотивации у пользователя
    required:
      - accountNumber
      - bikCode
      - bankName
    properties:
      accountNumber:
        type: string
        description: Номер счета банка РСХБ
        pattern: (^[0-9]{20})
      bikCode:
        type: string
        description: БИК банка
        pattern: (^[0-9]{9})
      bankName:  
        type: string
        description: Наименование филиала банка
        pattern: (^(.{0,250}$))
      inn:  
        type: string
        description: ИНН
        pattern: (^[0-9]{12})
      registrationAddress:  
        type: string
        description: Адрес регистрации, в случае отсутствия ИНН
        pattern: (^(.{0,250}$))
      index:  
        type: string
        description: Индекс, в случае отсутствия ИНН
        pattern: (^[0-9]{6})
      startDate:
        type: string
        format: date
        description: Дата начала действия програмы мотивации
  motivationProgramAdminList:
    type: object
    properties:
      elements:
        type: array
        items:
          $ref: '#/definitions/motivationProgramAdmin'
  motivationProgramAdmin:
    type: object
    description: Программа мотивации
    properties:
      id:
        type: integer
        format: int64
        description: Идентификатор мотивации
      userId:
        type: integer
        format: int64
        description: Идентификатор пользователя
      accountNumber:
        type: string
        description: Номер счета банка РСХБ
      bikCode:
        type: string
        description: БИК банка
      bankName:  
        type: string
        description: Наиминование филиала банка
      inn:  
        type: string
        description: ИНН
      registrationAddress:  
        type: string
        description: Адресс регистрации, в случае отсутствия ИНН 
      index:  
        type: string
        description: Индекс, в случае отсутствия ИНН
      comment:
        type: string
        description: комментарий
      motivationCorrectStatus:
        $ref: '#/definitions/motivationCorrectStatus'
      isActive:
        type: boolean
        description: Признак активности
      documentName:
        type: string
        description: Имя прикрепленного документа с плдтверждением участия
      passportName:
        type: string
        description: Паспорт участника программы мотивации
      startDate:
        type: string
        description: Дата начала действия програмы мотивации в формате ISO-8601 (ГГГГ-ММ-ДД)
      endDate:
        type: string
        description: Дата окончания действия програмы мотивации в формате ISO-8601 (ГГГГ-ММ-ДД)
  motivationSettingsEditAdmin:
    type: object
    description: Форма редактирования программы мотивации
    properties:
      accountNumber:
        type: string
        description: Номер счета банка РСХБ
        pattern: (^[0-9]{20})
      bikCode:
        type: string
        description: БИК банка
        pattern: (^[0-9]{9})
      bankName:
        type: string
        description: Наименование филиала банка
        pattern: (^(.{0,250}$))
      inn:
        type: string
        description: ИНН
        pattern: (^[0-9]{12})
      registrationAddress:
        type: string
        description: Адрес регистрации, в случае отсутствия ИНН
        pattern: (^(.{0,250}$))
      index:
        type: string
        description: Индекс, в случае отсутствия ИНН
        pattern: (^[0-9]{6})
      comment:
        type: string
        description: комментарий
      motivationCorrectStatus:
        $ref: '#/definitions/motivationCorrectStatus'
        description: Статус корректности мотивации
      startDate:
        type: string
        format: date
        description: Дата начала действия програмы мотивации
  motivationCorrectStatus:
    type: string
    description: >
      Статус корректности мотивации:
        CORRECT - Признак корректности
        INCORRECT - Признак некорректности
        NOT_CHECKED - Не проверено
    enum:
      - CORRECT
      - INCORRECT
      - NOT_CHECKED
...
  errorModel:
    type: object
    description: Описание ошибки
    required:
      - errors
    properties:
      errors:
        type: array
        items:
          $ref: '#/definitions/checkModel'
  checkModel:
    type: object
    description: Контейнер с результатами одной проверки
    properties:
      key:
        type: string
        description: Поле или иной объект, подверженный проверке
      value:
        type: string
        description: Результат проверки, описывающий выявленные отклонения от ТЗ
...